DCE-67: Refactor Search
