+++ src/engine/search/mod.rs
@@ -1,2 +1,5 @@
-pub mod searcher;
+pub mod alpha_beta;
+pub mod iter_deepening;
+pub mod pvs;
+pub mod quiescence;
 pub mod transposition_table;
