+++ src/engine/search/pvs.rs
@@ -1,0 +1,3 @@
+use super::iter_deepening::Search;
+
+impl Search {}
