+++ src/lib.rs
@@ -24,7 +24,10 @@ pub mod engine {
     }
 
     pub mod search {
-        pub mod searcher;
+        pub mod alpha_beta;
+        pub mod iter_deepening;
+        pub mod pvs;
+        pub mod quiescence;
         pub mod transposition_table;
     }
